% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerd_colors.R
\name{nerd_colors}
\alias{nerd_colors}
\title{Dark Nerd Themes Colors}
\usage{
nerd_colors(...)
}
\arguments{
\item{...}{Names of the colors}
}
\description{
Get nerd colors
}
\details{
Use \code{nerd_colors()}
}
\examples{
# Full list
nerd_colors()

# Choose colors
nerd_colors("chocolate")
nerd_colors("nerd-red", "nerd-blue")

}
