% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerd_colors.R
\name{nerd_pal}
\alias{nerd_pal}
\title{nerd color palettes generator.}
\usage{
nerd_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Returns a function helps pick a theme
}
