% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{nerd_palettes}
\alias{nerd_palettes}
\alias{nerd_main}
\alias{candy}
\alias{nord_frost}
\alias{nord_aurora}
\alias{everforest}
\title{nerd Color Palettes}
\usage{
nerd_main(n)

candy(n)

nord_frost(n)

nord_aurora(n)

everforest(n)
}
\arguments{
\item{n}{Number of colors required}
}
\description{
Functions to create color palettes from nerd theme colors
}
\details{
Color palettes built from popular dark nerd themes
}
\examples{
nerd_main(10)

}
