% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerd_theme.R
\name{nerd_theme_gruv}
\alias{nerd_theme_gruv}
\title{ggplot2 theme for nerd graphs.}
\usage{
nerd_theme_gruv(
  legend_right = FALSE,
  base_size = 12,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{legend_right}{Logical indicating whether legend should be placed to
the right of the plot. If FALSE, the default, legend is positioned above the
plot.}

\item{base_size}{The base font size}

\item{base_family}{Font family}

\item{base_line_size}{Default}

\item{base_rect_size}{Default}
}
\description{
ggplot2 theme for nerd graphs.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, hwy, color = class)) +
  geom_point() +
  nerd_theme()

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  nerd_theme(vars(class)) +
  nerd_theme()

}
